requirements.txt
pdf2image
pytesseract
faiss-cpu
sentence-transformers
numpy
